Fix some errors.
