+++ worlds/stardew_valley/data/items.csv
@@ -767,8 +767,10 @@ id,name,classification,groups,mod_name
 10123,Scarlett <3,progression,FRIENDSANITY,Stardew Valley Expanded
 10124,Susan <3,progression,FRIENDSANITY,Stardew Valley Expanded
 10125,Morris <3,progression,FRIENDSANITY,Stardew Valley Expanded
+10301,Progressive Woods Obelisk Sigils,progression,,DeepWoods
+10302,Progressive Skull Cavern Elevator,progression,,Skull Cavern Elevator
 10501,Marlon's Boat Paddle,progression,GINGER_ISLAND,Stardew Valley Expanded
-10502,Diamond Wand,useful,"GINGER_ISLAND",Stardew Valley Expanded
+10502,Diamond Wand,filler,"WEAPON,DEPRECATED",Stardew Valley Expanded
 10503,Iridium Bomb,progression,,Stardew Valley Expanded
 10504,Krobus' Protection,useful,,Stardew Valley Expanded
 10505,Kittyfish Spell,progression,,Stardew Valley Expanded
@@ -779,9 +781,7 @@ id,name,classification,groups,mod_name
 10510,Nexus: Outpost Runes,progression,MOD_WARP,Stardew Valley Expanded
 10511,Fable Reef Portal,progression,"GINGER_ISLAND,MOD_WARP",Stardew Valley Expanded
 10512,Super Starfruit,useful,,Stardew Valley Expanded
-10513,Tempered Galaxy Sword,progression,"TEMPERED_GALAXY_WEAPONS,WEAPON",Stardew Valley Expanded
-10514,Tempered Galaxy Dagger,progression,"TEMPERED_GALAXY_WEAPONS,WEAPON",Stardew Valley Expanded
-10515,Tempered Galaxy Hammer,progression,"TEMPERED_GALAXY_WEAPONS,WEAPON",Stardew Valley Expanded
+10513,Tempered Galaxy Sword,filler,"WEAPON,DEPRECATED",Stardew Valley Expanded
+10514,Tempered Galaxy Dagger,filler,"WEAPON,DEPRECATED",Stardew Valley Expanded
+10515,Tempered Galaxy Hammer,filler,"WEAPON,DEPRECATED",Stardew Valley Expanded
 10516,Abandoned House Outskirts Clean-up,progression,MOD_WARP,Stardew Valley Expanded
-10301,Progressive Woods Obelisk Sigils,progression,,DeepWoods
-10302,Progressive Skull Cavern Elevator,progression,,Skull Cavern Elevator
