+++ worlds/stardew_valley/logic/logic.py
@@ -577,6 +577,19 @@ class StardewLogic:
         traveling_merchant_days = [f"Traveling Merchant: {day}" for day in Weekday.all_days]
         return self.received(traveling_merchant_days, tier)
 
+    def can_grow_gold_quality(self, quality: int) -> StardewRule:
+        if quality <= 0:
+            return True_()
+        if quality == 1:
+            return self.skill.has_farming_level(5) | (self.crop.has_fertilizer(1) & self.skill.has_farming_level(2)) | (
+                    self.crop.has_fertilizer(2) & self.skill.has_farming_level(1)) | self.crop.has_fertilizer(3)
+        if quality == 2:
+            return self.skill.has_farming_level(10) | (self.crop.has_fertilizer(1) & self.skill.has_farming_level(5)) | (
+                    self.crop.has_fertilizer(2) & self.skill.has_farming_level(3)) | (
+                           self.crop.has_fertilizer(3) & self.skill.has_farming_level(2))
+        if quality >= 3:
+            return self.crop.has_fertilizer(3) & self.skill.has_farming_level(4)
+
     def can_complete_field_office(self) -> StardewRule:
         field_office = self.region.can_reach(Region.field_office)
         professor_snail = self.received("Open Professor Snail Cave")
